Revert "refactor: reader -> adpater 클래스명 수정"
This reverts commit 8ab327323b4601a2817bb103d9585cb8aec3f31b.
